connection refactoring
