+++ network/src/main/java/io/bitsquare/p2p/network/connection/InboundConnection.java
@@ -1,4 +1,4 @@
-package io.bitsquare.p2p.network;
+package io.bitsquare.p2p.network.connection;
 
 import java.net.Socket;
