+++ network/src/main/java/io/bitsquare/p2p/network/connection/OutboundConnection.java
@@ -1,4 +1,4 @@
-package io.bitsquare.p2p.network;
+package io.bitsquare.p2p.network.connection;
 
 import io.bitsquare.p2p.NodeAddress;
