+++ network/src/main/java/io/bitsquare/p2p/peers/PeerManager.java
@@ -6,6 +6,10 @@ import io.bitsquare.common.Timer;
 import io.bitsquare.common.UserThread;
 import io.bitsquare.p2p.NodeAddress;
 import io.bitsquare.p2p.network.*;
+import io.bitsquare.p2p.network.connection.CloseConnectionReason;
+import io.bitsquare.p2p.network.connection.Connection;
+import io.bitsquare.p2p.network.connection.ConnectionListener;
+import io.bitsquare.p2p.network.connection.InboundConnection;
 import io.bitsquare.p2p.peers.peerexchange.Peer;
 import io.bitsquare.storage.Storage;
 import org.slf4j.Logger;
