+++ network/src/main/java/io/bitsquare/p2p/peers/keepalive/KeepAliveHandler.java
@@ -7,8 +7,8 @@ import io.bitsquare.app.Log;
 import io.bitsquare.common.Timer;
 import io.bitsquare.common.UserThread;
 import io.bitsquare.p2p.Message;
-import io.bitsquare.p2p.network.Connection;
-import io.bitsquare.p2p.network.MessageListener;
+import io.bitsquare.p2p.network.connection.Connection;
+import io.bitsquare.p2p.network.connection.MessageListener;
 import io.bitsquare.p2p.network.NetworkNode;
 import io.bitsquare.p2p.peers.PeerManager;
 import io.bitsquare.p2p.peers.keepalive.messages.Ping;
