+++ network/src/test/java/io/bitsquare/p2p/network/TorNetworkNodeTest.java
@@ -5,6 +5,8 @@ import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.SettableFuture;
 import io.bitsquare.p2p.Message;
 import io.bitsquare.p2p.mocks.MockPayload;
+import io.bitsquare.p2p.network.connection.Connection;
+import io.bitsquare.p2p.network.connection.MessageListener;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.junit.Before;
 import org.junit.Ignore;
